apps/api/pyproject.toml
[project]
name = "ocr-api"
version = "0.1.0"
description = "OCR Platform API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "boto3>=1.35.0",
    "python-multipart>=0.0.9",
    "celery>=5.4.0",
    "python-dotenv>=1.0.0",
    "pypdf>=4.0.0",
    "ocr-core",
]

[tool.uv.sources]
ocr-core = { path = "../../libs/ocr_core" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["app"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = []

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
